requirements.txt
attrs
click
contoml==0.30
matplotlib
thoth-python
thoth-solver
thoth-storages
toml

